layout/style/test/test_shared_sheet_caching.html

<!doctype html>
<meta charset="utf-8">
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script>
  function childWindowLoaded(childWin) {
    is(
      childWin.getComputedStyle(childWin.document.documentElement).backgroundColor,
      "rgb(0, 255, 0)",
      "Sheet should apply",
    )
    is(
      SpecialPowers.getDOMWindowUtils(childWin).parsedStyleSheets,
      0,
      `Shouldn't need to parse the stylesheet ${childWin.parent == window ? "frame" : "top"}`
    );
    if (childWin.top == childWin) {
      SimpleTest.finish();
    }
  }
</script>
<link rel="stylesheet" href="file_shared_sheet_caching.css">
<iframe src="file_shared_sheet_caching.html"></iframe>
<a rel="opener" href="file_shared_sheet_caching.html" target="_blank">Navigation</a>
<script>
SimpleTest.waitForExplicitFinish();
onload = function() {
  info("Sheets parsed: " + SpecialPowers.DOMWindowUtils.parsedStyleSheets);
  is(
    getComputedStyle(document.documentElement).backgroundColor,
    "rgb(0, 255, 0)",
    "Sheet should apply",
  )
  document.querySelector("a").click();
}
</script>
